delete image when collection deleted

+++ src/routes/api/collection/[id].js
@@ -1,5 +1,6 @@
 import db from '../../../../db';
 import { responseBuilder } from '../_api';
+import { unlink } from "fs";
 
 export async function get({ params }) {
 	try {
@@ -12,7 +13,12 @@ export async function get({ params }) {
 
 export async function del({ params }) {
 	try {
+		const collection = await db.models.collection.findOne({ where: { id: params.id } });
+		if (!collection) return
 		await db.models.collection.destroy({ where: { id: params.id } });
+		unlink(`static${collection.imageUrl}`, (err) => {
+			console.log(err);
+		})
 		return responseBuilder(200, 'collection has ben deleted');
 	} catch (error) {
 		return responseBuilder(400, error);
